requirements.txt
# Travel Planning Agent - Dependencies
streamlit>=1.28.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
requests>=2.31.0
amadeus>=8.0.0
streamlit
google-generativeai
requests
python-dotenv

